+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed buffer handling in ContentCollections
+
 ## v0.14.6 (2022-11-28)
 * Switched to Cosmos for plugin interfaces
 * Migrated to use effigy in CI workflow